% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_or.R
\name{anonymise_count_values}
\alias{anonymise_count_values}
\title{Anonymise Count Values}
\usage{
anonymise_count_values(var)
}
\arguments{
\item{var}{A numeric vector containing count data.}
}
\value{
A character vector of anonymised counts.
}
\description{
Anonymises counts within a supplied numeric vector by applying specific formatting rules.
}
\details{
This function provides a vectorised approach to data anonymisation that:
\itemize{
\item Suppresses counts below 10 by replacing them with the string "<10".
\item Rounds counts of 10 or more to the nearest multiple of 5.
\item Formats counts using \code{scales::label_number_auto()} to include thousands separators, which respects the locale settings for choice of the big mark.
}

The function returns a vector of strings representing the anonymised counts.
}
