% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_or.R
\name{count_rows_by_variable}
\alias{count_rows_by_variable}
\title{Count Rows by Variable}
\usage{
count_rows_by_variable(df, var_name)
}
\arguments{
\item{df}{Tibble of data}

\item{var_name}{String name of the variable to count}
}
\value{
Tibble summarising the number of rows of data in var in total if var is numeric or by each level if var is a character
}
\description{
Takes a tibble of data and a string name of a variable in the tibble and returns a count of rows.
}
\details{
For numeric variables the result is a count of all non-NA rows.

For character and factor variables the result is a count of all non-NA rows per level / value of character.
}
