% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_or.R
\name{get_univariable_summary_table}
\alias{get_univariable_summary_table}
\title{Get a Summary Table of Univariable Analysis}
\usage{
get_univariable_summary_table(
  glm,
  conf_level = 0.95,
  confint_fast_estimate = FALSE,
  use_model_data_only = FALSE
)
}
\arguments{
\item{glm}{A binomial Generalised Linear Model (GLM) object from \code{\link[stats:glm]{stats::glm()}}.}

\item{conf_level}{Numeric value between 0.001 and 0.999 (default = 0.95) specifying the confidence level for the confidence interval.}

\item{confint_fast_estimate}{Boolean (default = \code{FALSE}) indicating whether to use a faster estimate of the confidence interval. Note: this assumes normally distributed data, which may not be suitable for your data.}

\item{use_model_data_only}{Boolean (default = \code{FALSE}) indicating whether to use only the subset of data that was used as part of the multivariable model, or set to \code{TRUE} to use the full set of data provided to the multivariable model. Note, any records (rows) containing missing values for any of the outcome or predictor variables is automatically excluded from the multivariable model by {stats::glm}, so the overall number of records used in multivariable models can be much lower than the total number of records supplied to the function. Set to \code{TRUE} to increase comparability between the univariable and multivariable models, set to \code{FALSE} to gain a more holistic view of the invididual relationships between predictors and outcome.}
}
\value{
Tibble providing a summary of the univariable logistic regression model.
}
\description{
Performs univariable logistic regression for each predictor in a
multivariable model, generating a summary table with odds ratios, confidence
intervals and p-values.
}
\details{
This function systematically:
\itemize{
\item Extracts predictors from a multivariable logistic regression model
\item Run individual logistic regression models for each predictor
\item Calculates odds ratios, confidence intervals and statistical significance
}

Key features:
\itemize{
\item Flexible confidence interval estimation
\item Option to use full or model-specific dataset
\item Comprehensive output for comparative analysis
}
}
\examples{
# prepare the data
df <-
  # get the dataset
  datasets::Titanic |>
  # convert to a tibble
  tibble::as_tibble() |>
  # convert the aggregate counts to individual observations
  tidyr::uncount(weights = n) |>
  # convert categorical variables to factors
  dplyr::mutate_if(
    .predicate = is.character,
    .funs = as.factor
  )

# create a model from the data
model <-
  stats::glm(
    formula = Survived ~ Age + Class + Sex,
    family = "binomial",
    data = df
  )

# get a univariable summary from all the data
plotor:::get_univariable_summary_table(
  glm =  model,
  confint_fast_estimate = TRUE
)
}
