% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_or.R
\name{label_groups}
\alias{label_groups}
\title{Labeller for groups}
\usage{
label_groups(group, level)
}
\arguments{
\item{group}{Character vector of groups}

\item{level}{Character vector of factor levels}
}
\description{
A custom labeller function for use in the OR plot.
}
\details{
It ensures factor groups are listed for the first instance of each level and as empty strings for any subsequent level. It is intended to reduce the visual clutter in the OR plot.
}
