% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_or.R
\name{plot_odds_ratio}
\alias{plot_odds_ratio}
\title{Plot the Odds Ratio}
\usage{
plot_odds_ratio(df, model)
}
\arguments{
\item{df}{Tibble of data containing a pre-prepared OR object}

\item{model}{A glm logistic regression model}
}
\value{
ggplot2 plot
}
\description{
Plot the OR plot using ggplot2 using points for OR estimates and whiskers for confidence intervals.
}
\details{
The plot is faceted on variables with sub-levels shown where the variable is a factor.
}
