% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotor-package.R
\docType{package}
\name{plotor-package}
\alias{plotor}
\alias{plotor-package}
\title{plotor: Odds Ratio Tools for Logistic Regression}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Produces odds ratio analyses with comprehensive reporting tools. Generates plots, summary tables, and diagnostic checks for logistic regression models fitted with 'glm()' using binomial family. Provides visualisation methods, formatted reporting tables via 'gt', and tools to assess logistic regression model assumptions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/craig-parylo/plotor}
  \item \url{https://craig-parylo.github.io/plotor/}
  \item Report bugs at \url{https://github.com/craig-parylo/plotor/issues}
}

}
\author{
\strong{Maintainer}: Craig Parylo \email{craig.parylo2@nhs.net} (\href{https://orcid.org/0000-0003-4297-7808}{ORCID}) [copyright holder]

}
\keyword{internal}
