% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_or.R
\name{prepare_df_for_plotting}
\alias{prepare_df_for_plotting}
\title{Prepare dataframe for plotting}
\usage{
prepare_df_for_plotting(df)
}
\arguments{
\item{df}{Tibble of data combining OR estimates and counts of rows per variables}
}
\value{
Tibble of data expanded with variables to aid plotting
}
\description{
Prepare dataframe for plotting
}
