% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_or.R
\name{summarise_rows_per_variable_in_model}
\alias{summarise_rows_per_variable_in_model}
\title{Get Summary of Rows per Variable in Model}
\usage{
summarise_rows_per_variable_in_model(model_results)
}
\arguments{
\item{model_results}{}
}
\value{
Tibble summary of rows per variable used in the model
}
\description{
Returns a summary of the number of rows per variable used in the model.
}
