% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_or.R
\name{use_var_labels}
\alias{use_var_labels}
\title{Use Variable Labels}
\usage{
use_var_labels(df, lr)
}
\arguments{
\item{df}{Tibble of data expanded with variables to aid plotting - as outputted from `prepare_df_for_plotting`}

\item{lr}{Results from a GLM binomial model results}
}
\value{
Tibble of data with group labels used where available
}
\description{
Where variables have been given a label attribute then the label is used in the plot
}
