% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_or.R
\name{validate_glm_model}
\alias{validate_glm_model}
\title{Validate the \code{{glm}} model}
\usage{
validate_glm_model(glm_model)
}
\arguments{
\item{glm_model}{Results from a binomial Generalised Linear Model (GLM), as produced by \code{\link[stats:glm]{stats::glm()}}}
}
\value{
boolean (TRUE = logistic regression, FALSE = other model)
}
\description{
Check whether the glm model object is the product of logistic regression.
}
