% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_or.R
\name{plot_or}
\alias{plot_or}
\title{Plot OR}
\usage{
plot_or(glm_model_results)
}
\arguments{
\item{glm_model_results}{Results from a GLM binomial model results}
}
\value{
ggplot2-produced visualisation
}
\description{
Produces an Odds Ratio plot to visualise the results of a logistic regression analysis.
}
