% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotor-package.R
\docType{package}
\name{plotor-package}
\alias{plotor}
\alias{plotor-package}
\title{plotor: Produces an Odds Ratio Plot from a Logistic Regression Model}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Produces an Odds Ratio (OR) Plot to visualise the result of a logistic regression analysis. Provide it with a binomial regression model produced by 'glm()' and it will convert the estimates to odds ratios with a 95% confidence interval and plot the results using 'ggplot2'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/craig-parylo/plotor}
  \item \url{https://craig-parylo.github.io/plotor/}
  \item Report bugs at \url{https://github.com/craig-parylo/plotor/issues}
}

}
\author{
\strong{Maintainer}: Craig Parylo \email{craig.parylo2@nhs.net} (\href{https://orcid.org/0000-0003-4297-7808}{ORCID}) [copyright holder]

}
\keyword{internal}
