% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotor-package.R
\docType{package}
\name{plotor-package}
\alias{plotor}
\alias{plotor-package}
\title{plotor: Produces an Odds Ratio Plot from a Logistic Regression Model}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Produces an odds‐ratio (OR) plot to visualise the results of a logistic regression model fitted with `glm(..., family = binomial)`. Estimates are converted to ORs with configurable confidence intervals (default is 95%). The package can return either a `ggplot2`-based plot, a formatted table via `gt`, or a set of diagnostic checks for model assumptions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/craig-parylo/plotor}
  \item \url{https://craig-parylo.github.io/plotor/}
  \item Report bugs at \url{https://github.com/craig-parylo/plotor/issues}
}

}
\author{
\strong{Maintainer}: Craig Parylo \email{craig.parylo2@nhs.net} (\href{https://orcid.org/0000-0003-4297-7808}{ORCID}) [copyright holder]

}
\keyword{internal}
